requirements.txt
-e .
aiostream==0.4.4
arrow==1.2.1
asyncio==3.4.3
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
py-algorand-sdk
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
tinyman-py-sdk @ git+https://github.com/tinymanorg/tinyman-py-sdk.git
tinyman-py-sdk-old @ git+https://github.com/lrnz-vtl/tinyman-py-sdk-old.git
urllib3==1.26.7
python-daemon==2.3.0
networkx==2.6.3
pyarrow==7.0.0
aiohttp==3.8.1
scikit-learn~=1.0.2
scipy~=1.8.0
statsmodels~=0.13.2
uvloop~=0.16.0
